hvac-pygame/requirements.txt
pygame>=2.5.0
paho-mqtt>=1.6.1
numpy>=1.24.0
